+++ src/components/molecules/index.ts
@@ -10,4 +10,3 @@ export { default as Pagination } from './Pagination'
 export { default as Modal } from './Modal'
 export { default as Tabs } from './Tabs'
 export { default as InputControl } from './InputControl'
-export { default as FormField } from './FormField'
